--- src/main/java/flcwGUI/GameController.java
@@ -1,16 +1,0 @@
-package flcwGUI;
-
-import javafx.fxml.FXML;
-import javafx.scene.control.Label;
-
-public class GameController {
-    @FXML
-    private Label welcomeText;
-
-    @FXML
-    protected void onHelloButtonClick() {
-        welcomeText.setText("Welcome to JavaFX Application!");
-    }
-
-
-}
