+++ src/main/java/flcwGUI/LaserChessGamePlay/chess/ChessOneWayMirror.java
@@ -12,6 +12,11 @@ public class ChessOneWayMirror extends Chess {
         this.ct = chess_type.OneWayMirror;
     }
 
+    @Override
+    public int getrotate() {
+        return direction.ordinal();
+    }
+
     //要求d的取值是r或者l
     public void rotate(char d){
         // 将枚举转换为数组
