+++ src/main/resources/flcwGUI/style.css
@@ -1,0 +1,57 @@
+/* styles.css */
+
+/* 设置场景的背景颜色 */
+.root {
+    -fx-background-image: url(../images/background.png);
+    -fx-background-size: cover;
+    -fx-opacity: 0.8;
+}
+
+/* 设置旋转按钮的样式 */
+.rotate-button {
+    -fx-background-color: #ffdd00; /* Green */
+    -fx-min-width:  100;
+    -fx-min-height: 100;
+    -fx-text-fill: white;
+    -fx-font-size: 15;
+}
+
+/* 设置旋转按钮容器的样式 */
+.rotate-button-container {
+    -fx-spacing: 60;
+    -fx-alignment: center;
+    -fx-padding: 0 0 60 0;
+}
+
+/* 设置退出按钮的样式 */
+.exit-button {
+    -fx-background-color: #fd3232; /* Red */
+    -fx-min-width:  60;
+    -fx-min-height: 50;
+    -fx-text-fill: #c2fff3;
+    -fx-font-size: 14;
+}
+
+.gameGrid {
+    -fx-border-color: #f8f2f2;
+    -fx-border-width: 10px;
+}
+
+/* 设置棋子的基础样式 */
+.chess {
+    -fx-min-width: 75;
+    -fx-min-height: 75;
+    -fx-background-radius: 20; /* 背景圆角 */
+    -fx-border-radius: 20; /* 边框圆角 */
+}
+
+/* 设置red_king样式 */
+.red_king {
+    -fx-background-image: url('../images/kings/red_king.png');
+}
+
+/* 设置blue_king样式 */
+.blue_king {
+    -fx-background-image: url('../images/kings/blue_king.png');
+}
+
